# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FairChat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages/{recipientId}:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages with a specific user
      parameters:
        - name: recipientId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: before
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Message history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message to a user
      parameters:
        - name: recipientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        recipientId:
          type: integer
        encryptedContent:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - senderId
        - recipientId
        - encryptedContent
        - createdAt
    SendMessageRequest:
      type: object
      properties:
        encryptedContent:
          type: string
      required:
        - encryptedContent
...