# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FairChat API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  senderId: number;
  recipientId: number;
  encryptedContent: string;
  createdAt: Date;
}